lab01

+++ Mikhail Bogdan/lab01/include/functions.h
@@ -2,7 +2,7 @@
 #define FUNCTIONS
 
 int doComputation(int a, float b);
-int max(int a, float b);
+int max(int a, int b);
 void logDebug(char message);
 
 #endif // !FUNCTIONS
